backend/requirements.txt
# Core
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.9
python-dotenv==1.1.1

# LLM clients
openai==0.28.0
anthropic==0.69.0
requests==2.32.5
PyYAML==6.0.3

# Rate limiting
slowapi==0.1.9

# Auth
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
